Rename files and format
