+++ src/commands/utility/delete-tag.js
